Create EditTextUtils so LoginActivity can use same keyboard logic
