+++ app/src/main/java/org/igarape/copcast/views/RegistrationActivity.java
@@ -18,6 +18,7 @@ import android.widget.TextView;
 import org.igarape.copcast.R;
 import org.igarape.copcast.service.sign.SigningService;
 import org.igarape.copcast.service.sign.SigningServiceException;
+import org.igarape.copcast.utils.EditTextUtils;
 import org.igarape.copcast.utils.Globals;
 import org.igarape.copcast.utils.OkDialog;
 import org.igarape.copcast.promises.Promise;
@@ -59,9 +60,9 @@ public class RegistrationActivity extends Activity {
             }
         });
 
-        showKeyboard(register_url);
-        showKeyboardOnFocusAndClick(register_url);
-        showKeyboardOnFocusAndClick(register_username);
+        EditTextUtils.showKeyboard(this, register_url);
+        EditTextUtils.showKeyboardOnFocusAndClick(this, register_url);
+        EditTextUtils.showKeyboardOnFocusAndClick(this, register_username);
     }
 
     private void doRegister() {
@@ -108,36 +109,5 @@ public class RegistrationActivity extends Activity {
                 return null;
             }
         }.execute();
-    }
-
-    private void showKeyboard(final TextView tv) {
-        tv.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                InputMethodManager keyboard = (InputMethodManager)
-                        getSystemService(Context.INPUT_METHOD_SERVICE);
-                keyboard.showSoftInput(tv, 0);
-            }
-        }, 200);
-    }
-
-    private void showKeyboardOnFocusAndClick(final TextView tv) {
-        // Show the keyboard when the user first focuses in the field.
-        tv.setOnFocusChangeListener(new View.OnFocusChangeListener() {
-            @Override
-            public void onFocusChange(View v, boolean hasFocus) {
-                if (hasFocus) {
-                    showKeyboard(tv);
-                }
-            }
-        });
-        // Show keyboard on click, in case the user has closed the keyboard.
-        // Click does not fire when the user first focuses in the field.
-        tv.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                showKeyboard(tv);
-            }
-        });
     }
 }
